# perf: improve isFileReadable performance