--- a/packages/vite/src/node/utils.ts
+++ b/packages/vite/src/node/utils.ts
@@ -537,7 +537,7 @@ export function writeFile(
 
 export function isFileReadable(filename: string): boolean {
   try {
-    // The "throwIfNoEntry" is performance optimization for cases where file does not exist
+    // The "throwIfNoEntry" is a performance optimization for cases where the file does not exist
     const fileExists = Boolean(fs.statSync(filename, { throwIfNoEntry: false }))
 
     if (!fileExists) {