### Description

In https://github.com/vitejs/vite/pull/6868 the `isFileReadable` was optimized to not use `try` + `catch` for figuring out if file does not exist. In https://github.com/vitejs/vite/pull/10793 this improvement was reverted in favor of filesystem access checks. This PR combines both of these together.

This change has huge impact in performance when large amount of filesystem reads are done. In https://github.com/vitest-dev/vitest/pull/3006 there is a reproduction case linked which runs 2.7s faster with these changes. If you need assistance for setting up the reproduction case let me know. 

#### Before:

```
 Test Files  5000 passed (5000)
      Tests  5000 passed (5000)
   Start at  13:54:37
   Duration  8.61s (transform 21.58s, setup 9ms, collect 26.79s, tests 4.05s)
```

The `captureLargerStackTrace` and `handleErrorFromBinding` are from [`fs.accessSync`](https://github.com/nodejs/node/blob/1ef4faa29bd5ef91462ce2dd1c3ec3e463d980b9/lib/fs.js#L248):

![image](https://user-images.githubusercontent.com/14806298/224699712-da55bc5b-a894-4b0a-8b16-792bd6bde38a.png)


#### After:

```
 Test Files  5000 passed (5000)
      Tests  5000 passed (5000)
   Start at  14:00:29
   Duration  5.73s (transform 10.47s, setup 9ms, collect 13.36s, tests 2.18s)
```

![image](https://user-images.githubusercontent.com/14806298/224699758-b0aaaac4-cf39-4904-8363-cf7b0656f199.png)


### Additional context

There are existing tests which nicely cover all requirements: https://github.com/vitejs/vite/blob/a595b115efbd38ac31c2de62ce5dd0faca424d02/packages/vite/src/node/__tests__/utils.spec.ts#L244-L267

The cost of using `try`+`catch` for checking file system is well explained here: https://marvinh.dev/blog/speeding-up-javascript-ecosystem-part-2/

---

### What is the purpose of this pull request? <!-- (put an "X" next to an item) -->

- [ ] Bug fix
- [ ] New Feature
- [ ] Documentation update
- [x] Other

### Before submitting the PR, please make sure you do the following

- [x] Read the [Contributing Guidelines](https://github.com/vitejs/vite/blob/main/CONTRIBUTING.md).
- [x] Read the [Pull Request Guidelines](https://github.com/vitejs/vite/blob/main/CONTRIBUTING.md#pull-request-guidelines) and follow the [PR Title Convention](https://github.com/vitejs/vite/blob/main/.github/commit-convention.md).
- [x] Check that there isn't already a PR that solves the problem the same way to avoid creating a duplicate.
- [x] Provide a description in this PR that addresses **what** the PR is solving, or reference the issue that it solves (e.g. `fixes #123`).
- [ ] Ideally, include relevant tests that fail without this PR but pass with it.