@AriPerkkio 2023-03-13
We just noticed that the `isFileReadable` seems to be not called that much on latest Vite. The performance metrics were recorded using `vite@4.0.0` which was included in the reproduction setup. The `isFileReadable` is identical in `4.0.0` and latest Vite, but is not used that much anymore. 

Anyway, I think it's still worth to merge this PR for possible future usage of `isFileReadable`.