@patak-dev 2023-03-13
For reference, @bluwy removed `isFileReadable` from the critical path in
- https://github.com/vitejs/vite/pull/11436

I still think we should merge this PR.