@patak-dev 2023-03-13
`isFileReadable` is used in the static and raw fs middleware, so even if it isn't in the critical path, `serveStaticMiddleware` comes before user post hooks (middlewares pushed in `configureServer`). The failing case is way slower than the positive case, so I think this optimization still makes sense. But I agree that we aren't in a hurry to merge it. Let's leave it open for a bit to see what others think.